Converting python values to double precision

+++ cython_opt.pyx
@@ -2,8 +2,8 @@ import numpy as np
 cimport numpy as np
 from cython_interface cimport python_to_c
 
-DTYPE = np.float32
-ctypedef np.float32_t DTYPE_t
+DTYPE = np.float64
+ctypedef np.float64_t DTYPE_t
 GPU_CHUNKS = 6
 
 def generate_indices(array_length, number_of_chunks):
@@ -36,7 +36,7 @@ def tensor_fit(ols_fit, design_matrix, signal, float min_signal, float min_diffu
     ols_columns = ols_fit.shape[1]
     design_rows = design_matrix.shape[0]
     design_columns = design_matrix.shape[1]
-    signal = signal.astype(float)
+    signal = signal.astype(double)
     signals = signal.shape[0]
     signal_elements = signal.shape[1]
